move set results to RaoData and adapt tests
