--- ra-optimisation/linear-rao/src/test/resources/LinearRaoParametersUnknownObjective.json
@@ -1,10 +1,0 @@
-{
-  "version" : "1.0",
-  "rao-with-loop-flow-limitation" : false,
-  "extensions" : {
-    "LinearRaoParameters" : {
-      "objective-function" : "UNKNOWN_OBJECTIVE",
-      "security-analysis-without-rao" : true
-    }
-  }
-}
