+++ ra-optimisation/rao-commons/src/main/java/com/farao_community/farao/rao_commons/linear_optimisation/core/fillers/MaxMinMarginFiller.java
@@ -10,7 +10,6 @@ package com.farao_community.farao.rao_commons.linear_optimisation.core.fillers;
 import com.farao_community.farao.commons.FaraoException;
 import com.farao_community.farao.data.crac_api.Cnec;
 import com.farao_community.farao.data.crac_api.PstRange;
-import com.farao_community.farao.data.crac_api.Unit;
 import com.farao_community.farao.rao_commons.RaoData;
 import com.farao_community.farao.rao_commons.linear_optimisation.core.LinearProblemParameters;
 import com.farao_community.farao.rao_commons.linear_optimisation.core.LinearProblem;
