+++ ra-optimisation/rao-commons/src/test/resources/LinearProblemParametersUnknownObjective.json
@@ -3,9 +3,7 @@
   "rao-with-loop-flow-limitation" : false,
   "extensions" : {
     "LinearProblemParameters" : {
-      "pst-penalty-cost" : 0.5,
-      "pst-sensitivity-threshold" : 1.0,
-      "objective-function" : "MAX_MIN_MARGIN_IN_AMPERE"
+      "objective-function" : "UNKNOWN"
     }
   }
 }
